Fixed hidpi issues in Win32
git-svn-id: svn://ultimatepp.org/upp/trunk@6086 f0d560ea-af0d-0410-9eb7-867de7ffcac7

+++ uppsrc/Core/FilterStream.cpp
@@ -120,7 +120,7 @@ void OutFilterStream::Close()
 void OutFilterStream::FlushOut()
 {
 	if(ptr != ~buffer) {
-		Filter(~buffer, ptr - ~buffer);
+		Filter(~buffer, (int)(ptr - ~buffer));
 		ptr = ~buffer;
 	}
 }
